Add files via upload
Place 'name.txt' and 'new.py' in a folder.
'name.txt' contains all vulnerable functions.
'h.c' is sample file in which we are checking.
Run 'new.py' : "for file in /home/priya/Documents/*; do python3 new.py "$file"; done"
A file 'name' will be created in the same folder where 'new.py' is present. It contains in which line which vulnerable function is present.